% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Dahir25.dat}
\alias{Dahir25.dat}
\title{Census-tract street-camera and crime data (Dahir 2025)}
\format{
A data frame with 11,620 rows and 19 columns:
\describe{
\item{city}{Factor with 10 U.S. city levels.}
\item{GEOID}{Census tract GEOID.}
\item{cam_count}{Integer count of street cameras in the tract
(the response used in the vignette).}
\item{image_count}{Integer count of street-view images sampled.}
\item{pnhwht}{Percent non-Hispanic white population (city-scaled).}
\item{pnhblk}{Percent non-Hispanic Black population (city-scaled).}
\item{pasian}{Percent Asian population.}
\item{phisp}{Percent Hispanic population.}
\item{entropy}{Race-ethnicity diversity entropy (city-scaled).}
\item{entropy_rank}{Within-city percentile rank of entropy.}
\item{total_crime_rate}{Total crime rate (city-scaled).}
\item{viol_veh_crime_rate}{Violent and vehicle crime rate.}
\item{pop}{Tract population (city-scaled).}
\item{hinc}{Median household income (city-scaled).}
\item{pvac}{Vacancy rate (city-scaled).}
\item{mhmval}{Median home value (city-scaled).}
\item{modal_zone}{Factor: dominant land-use zoning
(\code{commercial}, \code{industrial}, \code{mixed}, \code{public}, \code{residential},
\code{roads}).}
\item{road_length_km}{Tract road length in kilometres.}
\item{log_road_length}{Natural log of \code{road_length_km}, used as an
exposure offset.}
}
}
\source{
Replication data from Dahir et al. (2025).
}
\usage{
Dahir25.dat
}
\description{
Census-tract-level observations of CCTV / surveillance-camera counts paired
with sociodemographic and crime indicators from ten U.S. cities. Continuous
covariates are city-centred and scaled. Used in \pkg{glmOJ}'s vignette to
demonstrate count regression with an exposure offset
(\code{offset(log_road_length)}) and high-cardinality factor predictors.
}
\keyword{datasets}
